src/backend/requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
markdown==3.4.3
mysqlclient==2.1.1
django-cors-headers